requirements.txt
allure-python-commons~=2.13.5
pytest~=8.3.4
selenium~=4.27.1
webdriver-manager~=4.0.2
mysql-connector-python~=9.1.0
requests~=2.32.3